requirements.txt
asgiref==3.11.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
cryptography==46.0.5
Django==6.0.3
django-cors-headers==4.9.0
django-crontab==0.7.1
djangorestframework==3.16.1
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.11
oauthlib==3.3.1
pillow==12.1.1
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dotenv==1.2.2
qrcode==8.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
whitenoise==6.12.0
